+++ src/index.ts
@@ -1,12 +1,10 @@
 import * as dotenv from 'dotenv'
 import { httpServer } from './http_server/index'
-import { createWebSocketStream, WebSocketServer } from 'ws'
-import app from './app'
-import heartbeat from './helpers/heartbeat'
-import { TCommand, TWebSocket } from './types'
+import { WebSocketServer } from 'ws'
 import { interval } from './helpers/interval'
 import { resolve } from 'path'
 import { cwd } from 'process'
+import { connection } from './handlers/connection'
 
 dotenv.config({ path: resolve(cwd(), '.env') })
 
@@ -18,56 +16,7 @@ httpServer.listen(HTTP_PORT)
 
 export const wss = new WebSocketServer({ port: WSS_PORT })
 
-wss.on('connection', async (ws: TWebSocket) => {
-  ws.isAlive = true
-
-  ws.on('pong', heartbeat)
-
-  ws.on('close', () => {
-    duplex.destroy()
-  })
-
-  const duplex = createWebSocketStream(ws, {
-    encoding: 'utf8',
-    decodeStrings: false,
-  })
-
-  let data = ''
-
-  duplex.on('readable', async () => {
-    try {
-      let chunk
-
-      while (null !== (chunk = duplex.read())) {
-        data += chunk
-      }
-
-      const [command, ...params] = data.split(' ') as [TCommand]
-      const [x, y] = params.map(Number)
-
-      const runCommand = app()
-
-      const isValidCommand = Object.prototype.hasOwnProperty.call(
-        runCommand,
-        command
-      )
-
-      if (!isValidCommand) {
-        throw new Error(`${command} command not found`)
-      }
-
-      console.log(command, ...params)
-      const result = await runCommand[command](x, y)
-      duplex.write(`${command} ${result}`)
-    } catch (error) {
-      if (error instanceof Error) {
-        console.error(error.message)
-      }
-    } finally {
-      data = ''
-    }
-  })
-})
+wss.on('connection', connection())
 
 wss.on('close', () => {
   clearInterval(interval)
